Handle data cleaning, update tests, refactor naming.
