+++ pyhealth/datasets/__init__.py
@@ -55,7 +55,7 @@ from .covid19_cxr import COVID19CXRDataset
 from .dreamt import DREAMTDataset
 from .ehrshot import EHRShotDataset
 from .eicu import eICUDataset
-from .isic2018_bias import ISIC2018BiasDataset
+from .isic_bias import ISICBiasDataset
 from .isruc import ISRUCDataset
 from .medical_transcriptions import MedicalTranscriptionsDataset
 from .mimic3 import MIMIC3Dataset
