+++ pyhealth/datasets/configs/isic_bias.yaml
@@ -1,0 +1,31 @@
+version: "1.0"
+tables:
+    isic_artifacts_raw:
+      file_path: "isic_bias.csv"
+      sample_id: null
+      timestamp: null
+      attributes:
+        - "dark_corner"
+        - "hair"
+        - "gel_border"
+        - "gel_bubble"
+        - "ruler"
+        - "ink"
+        - "patches"
+        - "label"
+        - "label_string"
+    isic_artifacts_cleaned:
+      file_path: "isic_bias.csv"
+      sample_id: null
+      timestamp: null
+      attributes:
+        - "dark_corner"
+        - "hair"
+        - "gel_border"
+        - "gel_bubble"
+        - "ruler"
+        - "ink"
+        - "patches"
+        - "label"
+        - "label_string"
+
