+++ test-resources/core/isic_artifacts/isic_artifacts_cleaned/isic_bias.csv
@@ -1,4 +1,4 @@
-idx,sample_id,dark_corner,hair,gel_border,gel_bubble,ruler,ink,patches,label,label_string
+,image,dark_corner,hair,gel_border,gel_bubble,ruler,ink,patches,label,label_string
 0,ISIC_1,0,0,0,1,0,0,0,1,malignant
 1,ISIC_2,1,1,1,1,0,0,0,0,benign
 2,ISIC_3,0,1,0,1,0,0,0,0,malignant
